MET-93: Remove pagination from measurement families external endpoint
