--- src/Akeneo/Tool/Bundle/MeasureBundle/tests/EndToEnd/ExternalApi/Responses/measurement-families-count-limit.json
@@ -1,156 +1,0 @@
-{
-  "_links": {
-    "self": {
-      "href": "http://localhost/api/rest/v1/measurement-families?page=1&limit=1&with_count=true"
-    },
-    "first": {
-      "href": "http://localhost/api/rest/v1/measurement-families?page=1&limit=1&with_count=true"
-    },
-    "next": {
-      "href": "http://localhost/api/rest/v1/measurement-families?page=2&limit=1&with_count=true"
-    }
-  },
-  "current_page": 1,
-  "items_count": 23,
-  "_embedded": {
-    "items": [
-      {
-        "code": "Angle",
-        "labels": {
-          "de_DE": "Winkel",
-          "en_US": "Angle",
-          "fr_FR": "Angle"
-        },
-        "standard_unit_code": "RADIAN",
-        "units": [
-          {
-            "code": "RADIAN",
-            "labels": {
-              "en_US": "Radian",
-              "fr_FR": "Radian"
-            },
-            "convert_from_standard": [
-              {
-                "operator": "mul",
-                "value": "1"
-              }
-            ],
-            "symbol": "rad"
-          },
-          {
-            "code": "MILLIRADIAN",
-            "labels": {
-              "en_US": "Milliradian",
-              "fr_FR": "Milliradian"
-            },
-            "convert_from_standard": [
-              {
-                "operator": "mul",
-                "value": "0.001"
-              }
-            ],
-            "symbol": "mrad"
-          },
-          {
-            "code": "MICRORADIAN",
-            "labels": {
-              "en_US": "Microradian",
-              "fr_FR": "Microradian"
-            },
-            "convert_from_standard": [
-              {
-                "operator": "mul",
-                "value": "1.0E-6"
-              }
-            ],
-            "symbol": "µrad"
-          },
-          {
-            "code": "DEGREE",
-            "labels": {
-              "en_US": "Degree",
-              "fr_FR": "Degré"
-            },
-            "convert_from_standard": [
-              {
-                "operator": "mul",
-                "value": "0.01745329"
-              }
-            ],
-            "symbol": "°"
-          },
-          {
-            "code": "MINUTE",
-            "labels": {
-              "en_US": "Minute",
-              "fr_FR": "Minute"
-            },
-            "convert_from_standard": [
-              {
-                "operator": "mul",
-                "value": "0.0002908882"
-              }
-            ],
-            "symbol": "'"
-          },
-          {
-            "code": "SECOND",
-            "labels": {
-              "en_US": "Second",
-              "fr_FR": "Seconde"
-            },
-            "convert_from_standard": [
-              {
-                "operator": "mul",
-                "value": "4.848137E-6"
-              }
-            ],
-            "symbol": "\""
-          },
-          {
-            "code": "GON",
-            "labels": {
-              "en_US": "Gon",
-              "fr_FR": "Gon"
-            },
-            "convert_from_standard": [
-              {
-                "operator": "mul",
-                "value": "0.01570796"
-              }
-            ],
-            "symbol": "gon"
-          },
-          {
-            "code": "MIL",
-            "labels": {
-              "en_US": "Mil",
-              "fr_FR": "Mil"
-            },
-            "convert_from_standard": [
-              {
-                "operator": "mul",
-                "value": "0.0009817477"
-              }
-            ],
-            "symbol": "mil"
-          },
-          {
-            "code": "REVOLUTION",
-            "labels": {
-              "en_US": "Revolution",
-              "fr_FR": "Révolution"
-            },
-            "convert_from_standard": [
-              {
-                "operator": "mul",
-                "value": "6.283185"
-              }
-            ],
-            "symbol": "rev"
-          }
-        ]
-      }
-    ]
-  }
-}
