--- src/Akeneo/Tool/Bundle/MeasureBundle/tests/EndToEnd/ExternalApi/Responses/measurement-families-page-limit.json
@@ -1,245 +1,0 @@
-{
-  "_links": {
-    "self": {
-      "href": "http://localhost/api/rest/v1/measurement-families?page=2&limit=3&with_count=false"
-    },
-    "first": {
-      "href": "http://localhost/api/rest/v1/measurement-families?page=1&limit=3&with_count=false"
-    },
-    "previous": {
-      "href": "http://localhost/api/rest/v1/measurement-families?page=1&limit=3&with_count=false"
-    },
-    "next": {
-      "href": "http://localhost/api/rest/v1/measurement-families?page=3&limit=3&with_count=false"
-    }
-  },
-  "current_page": 2,
-  "_embedded": {
-    "items": [
-      {
-        "code": "Brightness",
-        "labels": {
-          "de_DE": "Helligkeit",
-          "en_US": "Brightness",
-          "fr_FR": "Luminosité"
-        },
-        "standard_unit_code": "LUMEN",
-        "units": [
-          {
-            "code": "LUMEN",
-            "labels": {
-              "de_DE": "Lumen",
-              "en_US": "Lumen",
-              "fr_FR": "Lumen"
-            },
-            "convert_from_standard": [
-              {
-                "operator": "mul",
-                "value": "1"
-              }
-            ],
-            "symbol": "lm"
-          },
-          {
-            "code": "NIT",
-            "labels": {
-              "de_DE": "Nit",
-              "en_US": "Nit",
-              "fr_FR": "Nit"
-            },
-            "convert_from_standard": [
-              {
-                "operator": "mul",
-                "value": "0.2918855809"
-              }
-            ],
-            "symbol": "nits"
-          }
-        ]
-      },
-      {
-        "code": "Capacitance",
-        "labels": {
-          "de_DE": "Elektrische Kapazität",
-          "en_US": "Capacitance",
-          "fr_FR": "Capacité électrique"
-        },
-        "standard_unit_code": "FARAD",
-        "units": [
-          {
-            "code": "ATTOFARAD",
-            "labels": {
-              "en_US": "Attofarad",
-              "fr_FR": "Attofarad"
-            },
-            "convert_from_standard": [
-              {
-                "operator": "div",
-                "value": "1000000000000000000"
-              }
-            ],
-            "symbol": "aF"
-          },
-          {
-            "code": "PICOFARAD",
-            "labels": {
-              "en_US": "Picofarad",
-              "fr_FR": "Picofarad"
-            },
-            "convert_from_standard": [
-              {
-                "operator": "div",
-                "value": "1000000000000"
-              }
-            ],
-            "symbol": "pF"
-          },
-          {
-            "code": "NANOFARAD",
-            "labels": {
-              "en_US": "Nanofarad",
-              "fr_FR": "Nanofarad"
-            },
-            "convert_from_standard": [
-              {
-                "operator": "div",
-                "value": "1000000000"
-              }
-            ],
-            "symbol": "nF"
-          },
-          {
-            "code": "MICROFARAD",
-            "labels": {
-              "en_US": "Microfarad",
-              "fr_FR": "Microfarad"
-            },
-            "convert_from_standard": [
-              {
-                "operator": "div",
-                "value": "1000000"
-              }
-            ],
-            "symbol": "µF"
-          },
-          {
-            "code": "MILLIFARAD",
-            "labels": {
-              "en_US": "Millifarad",
-              "fr_FR": "Millifarad"
-            },
-            "convert_from_standard": [
-              {
-                "operator": "div",
-                "value": "1000"
-              }
-            ],
-            "symbol": "mF"
-          },
-          {
-            "code": "FARAD",
-            "labels": {
-              "en_US": "Farad",
-              "fr_FR": "Farad"
-            },
-            "convert_from_standard": [
-              {
-                "operator": "mul",
-                "value": "1"
-              }
-            ],
-            "symbol": "F"
-          },
-          {
-            "code": "KILOFARAD",
-            "labels": {
-              "en_US": "Kilofarad",
-              "fr_FR": "Kilofarad"
-            },
-            "convert_from_standard": [
-              {
-                "operator": "mul",
-                "value": "1000"
-              }
-            ],
-            "symbol": "kF"
-          }
-        ]
-      },
-      {
-        "code": "CaseBox",
-        "labels": {
-          "ca_ES": "Embalatge",
-          "da_DK": "Emballage",
-          "de_DE": "Verpackung",
-          "en_GB": "Packaging",
-          "en_NZ": "Packaging",
-          "en_US": "Packaging",
-          "es_ES": "Embalaje",
-          "fi_FI": "Pakkaus",
-          "fr_FR": "Emballage",
-          "it_IT": "Imballo",
-          "ja_JP": "包装",
-          "pt_BR": "Embalagens",
-          "ru_RU": "Упаковка",
-          "sv_SE": "Förpackning"
-        },
-        "standard_unit_code": "PIECE",
-        "units": [
-          {
-            "code": "PIECE",
-            "labels": {
-              "ca_ES": "Peça",
-              "da_DK": "Stykke",
-              "de_DE": "Stück",
-              "en_GB": "Piece",
-              "en_NZ": "Piece",
-              "en_US": "Piece",
-              "es_ES": "Pieza",
-              "fi_FI": "Kappale",
-              "fr_FR": "Unité",
-              "it_IT": "Pezzo",
-              "ja_JP": "作品",
-              "pt_BR": "Peça",
-              "ru_RU": "шт.",
-              "sv_SE": "Stycke"
-            },
-            "convert_from_standard": [
-              {
-                "operator": "mul",
-                "value": "1"
-              }
-            ],
-            "symbol": "Pc"
-          },
-          {
-            "code": "DOZEN",
-            "labels": {
-              "ca_ES": "Dotzena",
-              "da_DK": "Dusin",
-              "de_DE": "Dutzend",
-              "en_GB": "Dozen",
-              "en_NZ": "Dozen",
-              "en_US": "Dozen",
-              "es_ES": "Docena",
-              "fi_FI": "Tusina",
-              "fr_FR": "Douzaine",
-              "it_IT": "Dozzina",
-              "ja_JP": "ダース",
-              "pt_BR": "Dúzia",
-              "ru_RU": "Дюжина",
-              "sv_SE": "Dussin"
-            },
-            "convert_from_standard": [
-              {
-                "operator": "mul",
-                "value": "12"
-              }
-            ],
-            "symbol": "Dz"
-          }
-        ]
-      }
-    ]
-  }
-}
